pyproject.toml
    [project]
    name = "laptop-screen-editor"
    version = "0.1.0"
    readme = "README.md"
    packages = [{ include = "app" }]
    requires-python = ">=3.11,<3.12"

    [tool.poetry.dependencies]
    python = ">=3.11,<3.12"

    # --- Computer Vision ---
    opencv-python = "^4.12.0.88"
    numpy = ">=2.0,<2.3.0"

    # --- Image IO ---
    pillow = "^10.4.0"
    pillow-heif = "^0.18.0"

    # --- UI ---
    PySide6 = "^6.7.2"

    # --- Utils ---
    loguru = "^0.7.3"
    tqdm = "^4.67.1"

    # --- Config / Validation ---
    pydantic = "^2.10.5"
    pydantic-settings = "^2.7.0"
    ultralytics = "^8.3.246"
    pillow-avif-plugin = "^1.5.2"

    [tool.poetry.group.dev.dependencies]
    pytest = "^8.3.4"
    pytest-qt = "^4.4.0"
    black = "^24.10.0"
    ruff = "^0.8.4"
    pyinstaller = "^6.17.0"

                            [tool.poetry.group.ml]
    optional = true

    [tool.poetry.group.ml.dependencies]
    ultralytics = "^8.3.0"
    onnxruntime = "^1.20.1"

    torch = { version = "^2.5.0", optional = true }
    torchvision = { version = "^0.20.0", optional = true }

    [build-system]
    requires = ["poetry-core"]
    build-backend = "poetry.core.masonry.api"

    [tool.black]
    line-length = 100
    target-version = ["py311"]

    [tool.ruff]
    line-length = 100
    target-version = "py311"

    [tool.pytest.ini_options]
    testpaths = ["tests"]
    python_files = ["test_*.py"]
